Adding Playlist Related Use Cases To the project

+++ app/src/main/java/com/soroush/eskandarie/musicplayer/presentation/ui/model/PlaylistDropdownItem.kt
@@ -5,5 +5,5 @@ import android.net.Uri
 data class PlaylistDropdownItem(
     val id:     Int,
     val name:   String,
-    val onClick: (id: Int) -> Unit
+    val onClick: (id: Long) -> Unit
 )
